[Refactor] Separate runtime from build_module
